unstupid

+++ src/server/ClientsConnection.cxx
@@ -72,7 +72,7 @@ void ClientsConnection::listen(){
 
 }
 
-void ClientsConnection::sendPacket(Packet *p, uint32_t client)
+void ClientsConnection::sendPacket(Packet *p, uint64_t client)
 {
 	nm_mutex.lock();
 		for(int i = 0; i < packetTransporters.size(); i++)
